added initial physical drawings
few changes to the original. first, the right half will not have 2 columns of shelves for the white boxes. instead it will have one column of shelves. another change, the bottom row of drawers on the right will be 3.5" tall and the top two rows of drawers on the right will be 3.25" tall.